refactor(vault): migrate ProfileManager & CategoryManager to feature-based architecture
