+++ frontend/src/features/vault/components/index.ts
@@ -1,0 +1,15 @@
+// Barrel export for vault components
+export { default as ProfileManager } from './ProfileManager';
+export { default as ProfileList } from './ProfileList';
+export { default as ProfileCard } from './ProfileCard';
+export { default as CreditCardItem } from './CreditCardItem';
+export { default as CredentialField } from './CredentialField';
+
+// Category components
+export { default as CategoryManager } from './CategoryManager';
+export { default as CategorySection } from './CategorySection';
+export { default as OrganizationCard } from './OrganizationCard';
+export { default as DigitalWalletGrid } from './DigitalWalletGrid';
+
+// Re-export card components
+export * from './cards';
