+++ frontend/src/features/vault/hooks/index.ts
@@ -3,8 +3,5 @@
  * Vault Hooks Barrel Export
  */
 
-// Hooks will be added here as they are created
-// Example: export { useVault } from './useVault';
-
-// Empty export to make this a valid module
-export {};
+export { useProfiles } from './useProfiles';
+export { useCategories } from './useCategories';
